# Optimize and fix integer overflow in subsecond parsing.