--- a/Data/Aeson/Parser/Time.hs
+++ b/Data/Aeson/Parser/Time.hs
@@ -20,7 +20,7 @@ module Data.Aeson.Parser.Time
     , zonedTime
     ) where
 
-import Control.Applicative ((<$>), (<|>), empty)
+import Control.Applicative ((<$>), (<*>), (<*), (*>))
 import Control.Monad (when, void)
 import Data.Attoparsec.Text as A
 import Data.Bits ((.&.))