@lpsmith 2015-07-18
What I'm finding a bit mysterious is why this change adds about a microsecond to the benchmarks that don't contain seconds,  and why this change saves ~6-7 microseconds when parsing ZonedTime but only ~3-4 microseconds when parsing UTCTime.    This result appears to be fairly consistent across repeated runs on my laptop.