@bos 2015-07-20
Nice work. I applied your change, and I was able to make it faster still via f899572201f87643da58cba7093b5a2b201b1610 :-)